unsigned integer - reversed approach

+++ src/map.cpp
@@ -4578,13 +4578,13 @@ void map::make_active( item_location &loc )
 // date to current time, and also check contents.
 void map::apply_in_fridge( item &it, int temp )
 {
-    int diff_freeze = abs(temp - FREEZING_TEMPERATURE);
-    diff_freeze = std::max( 1, diff_freeze );
-    diff_freeze = std::min( 5, diff_freeze );
+    unsigned int diff_freeze = abs(temp - FREEZING_TEMPERATURE);
+    diff_freeze = std::max( static_cast<unsigned int>(1), diff_freeze );
+    diff_freeze = std::min( static_cast<unsigned int>(5), diff_freeze );
 
-    int diff_cold = abs(temp - FRIDGE_TEMPERATURE);
-    diff_freeze = std::max( 1, diff_cold );
-    diff_freeze = std::min( 5, diff_cold );
+    unsigned int diff_cold = abs(temp - FRIDGE_TEMPERATURE);
+    diff_freeze = std::max( static_cast<unsigned int>(1), diff_cold );
+    diff_freeze = std::min( static_cast<unsigned int>(5), diff_cold );
 
     if( it.is_food() ) {
         if( temp <= FREEZING_TEMPERATURE ) {
